updated for fixes to documents for blob and proper typescript support
