+++ example/ios/App/App/public/assets/ios.transition-legacy-D1KhzRJI.js
@@ -1,0 +1,5 @@
+!function(){function t(e){return t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t},t(e)}function e(e,o,n){return(o=function(e){var o=function(e,o){if("object"!=t(e)||!e)return e;var n=e[Symbol.toPrimitive];if(void 0!==n){var r=n.call(e,o||"default");if("object"!=t(r))return r;throw new TypeError("@@toPrimitive must return a primitive value.")}return("string"===o?String:Number)(e)}(e,"string");return"symbol"==t(o)?o:o+""}(o))in e?Object.defineProperty(e,o,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[o]=n,e}function o(t,e){var o="undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(!o){if(Array.isArray(t)||(o=function(t,e){if(t){if("string"==typeof t)return n(t,e);var o={}.toString.call(t).slice(8,-1);return"Object"===o&&t.constructor&&(o=t.constructor.name),"Map"===o||"Set"===o?Array.from(t):"Arguments"===o||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(o)?n(t,e):void 0}}(t))||e&&t&&"number"==typeof t.length){o&&(t=o);var r=0,a=function(){};return{s:a,n:function(){return r>=t.length?{done:!0}:{done:!1,value:t[r++]}},e:function(t){throw t},f:a}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var l,i=!0,c=!1;return{s:function(){o=o.call(t)},n:function(){var t=o.next();return i=t.done,t},e:function(t){c=!0,l=t},f:function(){try{i||null==o.return||o.return()}finally{if(c)throw l}}}}function n(t,e){(null==e||e>t.length)&&(e=t.length);for(var o=0,n=Array(e);o<e;o++)n[o]=t[o];return n}System.register(["./index-legacy-CRsE_5AH.js"],(function(t,n){"use strict";var r,a;return{setters:[function(t){r=t.n,a=t.o}],execute:function(){
+/*!
+             * (C) Ionic http://ionicframework.com - MIT License
+             */
+var n=function(t){return document.querySelector("".concat(t,".ion-cloned-element"))},l=t("shadow",(function(t){return t.shadowRoot||t})),i=function(t){var e="ION-TABS"===t.tagName?t:t.querySelector("ion-tabs"),o="ion-content ion-header:not(.header-collapse-condense-inactive) ion-title.title-large";if(null!=e){var n=e.querySelector("ion-tab:not(.tab-hidden), .ion-page:not(.ion-page-hidden)");return null!=n?n.querySelector(o):null}return t.querySelector(o)},c=function(t,e){var n="ION-TABS"===t.tagName?t:t.querySelector("ion-tabs"),r=[];if(null!=n){var a=n.querySelector("ion-tab:not(.tab-hidden), .ion-page:not(.ion-page-hidden)");null!=a&&(r=a.querySelectorAll("ion-buttons"))}else r=t.querySelectorAll("ion-buttons");var l,i=o(r);try{for(i.s();!(l=i.n()).done;){var c=l.value,s=c.closest("ion-header"),f=s&&!s.classList.contains("header-collapse-condense-inactive"),d=c.querySelector("ion-back-button"),u=c.classList.contains("buttons-collapse"),m="start"===c.slot||""===c.slot;if(null!==d&&m&&(u&&f&&e||!u))return d}}catch(y){i.e(y)}finally{i.f()}return null},s=function(t,o,a,i,c,s,f,u,m){var y,p,b=o?"calc(100% - ".concat(c.right+4,"px)"):"".concat(c.left-4,"px"),v=o?"right":"left",g=o?"left":"right",S=o?"right":"left",h=1,T=1,x="scale(".concat(T,")"),q="scale(1)";if(s&&f){var A=(null===(y=s.textContent)||void 0===y?void 0:y.trim())===(null===(p=u.textContent)||void 0===p?void 0:p.trim());h=m.width/f.width,T=(m.height-d)/f.height,x=A?"scale(".concat(h,", ").concat(T,")"):"scale(".concat(T,")")}var E=l(i).querySelector("ion-icon").getBoundingClientRect(),X="".concat(o?E.width/2-(E.right-c.right):c.left-E.width/2,"px"),w=o?"-".concat(window.innerWidth-c.right,"px"):"".concat(c.left,"px"),C="".concat(m.top,"px"),k="".concat(c.top,"px"),P=[{offset:0,transform:"translate3d(".concat(X,", ").concat(C,", 0)")},{offset:1,transform:"translate3d(".concat(w,", ").concat(k,", 0)")}],B=[{offset:0,transform:"translate3d(".concat(w,", ").concat(k,", 0)")},{offset:1,transform:"translate3d(".concat(X,", ").concat(C,", 0)")}],R=a?B:P,W=a?[{offset:0,opacity:1,transform:q},{offset:1,opacity:0,transform:x}]:[{offset:0,opacity:0,transform:x},{offset:1,opacity:1,transform:q}],L=a?[{offset:0,opacity:1,transform:"scale(1)"},{offset:.2,opacity:0,transform:"scale(0.6)"},{offset:1,opacity:0,transform:"scale(0.6)"}]:[{offset:0,opacity:0,transform:"scale(0.6)"},{offset:.6,opacity:0,transform:"scale(0.6)"},{offset:1,opacity:1,transform:"scale(1)"}],j=r(),I=r(),N=r(),O=n("ion-back-button"),z=l(O).querySelector(".button-text"),D=l(O).querySelector("ion-icon");O.text=i.text,O.mode=i.mode,O.icon=i.icon,O.color=i.color,O.disabled=i.disabled,O.style.setProperty("display","block"),O.style.setProperty("position","fixed"),I.addElement(D),j.addElement(z),N.addElement(O),N.beforeStyles(e({position:"absolute",top:"0px"},S,"0px")).beforeAddWrite((function(){i.style.setProperty("display","none"),O.style.setProperty(v,b)})).afterAddWrite((function(){i.style.setProperty("display",""),O.style.setProperty("display","none"),O.style.removeProperty(v)})).keyframes(R),j.beforeStyles({"transform-origin":"".concat(v," top")}).keyframes(W),I.beforeStyles({"transform-origin":"".concat(g," center")}).keyframes(L),t.addAnimation([j,I,N])},f=function(t,o,a,l,i,c,s,f,u){var m,y,p=o?"right":"left",b=o?"calc(100% - ".concat(i.right,"px)"):"".concat(i.left,"px"),v="".concat(i.top,"px"),g=o?"-".concat(window.innerWidth-s.right-8,"px"):"".concat(s.x+8,"px"),S=.5,h="scale(1)",T="scale(".concat(S,")");if(f&&u){g=o?"-".concat(window.innerWidth-u.right-8,"px"):"".concat(u.x-8,"px");var x=(null===(m=f.textContent)||void 0===m?void 0:m.trim())===(null===(y=l.textContent)||void 0===y?void 0:y.trim()),q=u.width/c.width;S=u.height/(c.height-d),T=x?"scale(".concat(q,", ").concat(S,")"):"scale(".concat(S,")")}var A=s.top+s.height/2,E=i.height*S/2,X="".concat(A-E,"px"),w=[{offset:0,opacity:0,transform:"translate3d(".concat(g,", ").concat(X,", 0) ").concat(T)},{offset:.1,opacity:0},{offset:1,opacity:1,transform:"translate3d(".concat("0px",", ").concat(v,", 0) ").concat(h)}],C=[{offset:0,opacity:.99,transform:"translate3d(".concat("0px",", ").concat(v,", 0) ").concat(h)},{offset:.6,opacity:0},{offset:1,opacity:0,transform:"translate3d(".concat(g,", ").concat(X,", 0) ").concat(T)}],k=a?w:C,P=n("ion-title"),B=r();P.innerText=l.innerText,P.size=l.size,P.color=l.color,B.addElement(P),B.beforeStyles(e({"transform-origin":"".concat(p," top"),height:"".concat(i.height,"px"),display:"",position:"relative"},p,b)).beforeAddWrite((function(){l.style.setProperty("opacity","0")})).afterAddWrite((function(){l.style.setProperty("opacity",""),P.style.setProperty("display","none")})).keyframes(k),t.addAnimation(B)},d=(t("iosTransitionAnimation",(function(t,e){var o;try{var n="opacity",d="transform",u="0%",m="rtl"===t.ownerDocument.dir,y=m?"-99.5%":"99.5%",p=m?"33%":"-33%",b=e.enteringEl,v=e.leavingEl,g="back"===e.direction,S=b.querySelector(":scope > ion-content"),h=b.querySelectorAll(":scope > ion-header > *:not(ion-toolbar), :scope > ion-footer > *"),T=b.querySelectorAll(":scope > ion-header > ion-toolbar"),x=r(),q=r();if(x.addElement(b).duration((null!==(o=e.duration)&&void 0!==o?o:0)||540).easing(e.easing||"cubic-bezier(0.32,0.72,0,1)").fill("both").beforeRemoveClass("ion-page-invisible"),v&&null!=t){var A=r();A.addElement(t),x.addAnimation(A)}if(S||0!==T.length||0!==h.length?(q.addElement(S),q.addElement(h)):q.addElement(b.querySelector(":scope > .ion-page, :scope > ion-nav, :scope > ion-tabs")),x.addAnimation(q),g?q.beforeClearStyles([n]).fromTo("transform","translateX(".concat(p,")"),"translateX(".concat(u,")")).fromTo(n,.8,1):q.beforeClearStyles([n]).fromTo("transform","translateX(".concat(y,")"),"translateX(".concat(u,")")),S){var E=l(S).querySelector(".transition-effect");if(E){var X=E.querySelector(".transition-cover"),w=E.querySelector(".transition-shadow"),C=r(),k=r(),P=r();C.addElement(E).beforeStyles({opacity:"1",display:"block"}).afterStyles({opacity:"",display:""}),k.addElement(X).beforeClearStyles([n]).fromTo(n,0,.1),P.addElement(w).beforeClearStyles([n]).fromTo(n,.03,.7),C.addAnimation([k,P]),q.addAnimation([C])}}var B=b.querySelector("ion-header.header-collapse-condense"),R=function(t,e,o,n,r){var a=c(n,o),d=i(r),u=i(n),m=c(r,o),y=null!==a&&null!==d&&!o,p=null!==u&&null!==m&&o;if(y){var b=d.getBoundingClientRect(),v=a.getBoundingClientRect(),g=l(a).querySelector(".button-text"),S=null==g?void 0:g.getBoundingClientRect(),h=l(d).querySelector(".toolbar-title").getBoundingClientRect();f(t,e,o,d,b,h,v,g,S),s(t,e,o,a,v,g,S,d,h)}else if(p){var T=u.getBoundingClientRect(),x=m.getBoundingClientRect(),q=l(m).querySelector(".button-text"),A=null==q?void 0:q.getBoundingClientRect(),E=l(u).querySelector(".toolbar-title").getBoundingClientRect();f(t,e,o,u,T,E,x,q,A),s(t,e,o,m,x,q,A,u,E)}return{forward:y,backward:p}}(x,m,g,b,v),W=R.forward,L=R.backward;if(T.forEach((function(t){var e=r();e.addElement(t),x.addAnimation(e);var o=r();o.addElement(t.querySelector("ion-title"));var a,i=r(),c=Array.from(t.querySelectorAll("ion-buttons,[menuToggle]")),s=t.closest("ion-header"),f=null==s?void 0:s.classList.contains("header-collapse-condense-inactive");a=g?c.filter((function(t){var e=t.classList.contains("buttons-collapse");return e&&!f||!e})):c.filter((function(t){return!t.classList.contains("buttons-collapse")})),i.addElement(a);var d=r();d.addElement(t.querySelectorAll(":scope > *:not(ion-title):not(ion-buttons):not([menuToggle])"));var b=r();b.addElement(l(t).querySelector(".toolbar-background"));var v=r(),S=t.querySelector("ion-back-button");if(S&&v.addElement(S),e.addAnimation([o,i,d,b,v]),i.fromTo(n,.01,1),d.fromTo(n,.01,1),g)f||o.fromTo("transform","translateX(".concat(p,")"),"translateX(".concat(u,")")).fromTo(n,.01,1),d.fromTo("transform","translateX(".concat(p,")"),"translateX(".concat(u,")")),v.fromTo(n,.01,1);else if(B||o.fromTo("transform","translateX(".concat(y,")"),"translateX(".concat(u,")")).fromTo(n,.01,1),d.fromTo("transform","translateX(".concat(y,")"),"translateX(".concat(u,")")),b.beforeClearStyles([n,"transform"]),(null==s?void 0:s.translucent)?b.fromTo("transform",m?"translateX(-100%)":"translateX(100%)","translateX(0px)"):b.fromTo(n,.01,"var(--opacity)"),W||v.fromTo(n,.01,1),S&&!W){var h=r();h.addElement(l(S).querySelector(".button-text")).fromTo("transform",m?"translateX(-100px)":"translateX(100px)","translateX(0px)"),e.addAnimation(h)}})),v){var j=r(),I=v.querySelector(":scope > ion-content"),N=v.querySelectorAll(":scope > ion-header > ion-toolbar"),O=v.querySelectorAll(":scope > ion-header > *:not(ion-toolbar), :scope > ion-footer > *");if(I||0!==N.length||0!==O.length?(j.addElement(I),j.addElement(O)):j.addElement(v.querySelector(":scope > .ion-page, :scope > ion-nav, :scope > ion-tabs")),x.addAnimation(j),g){j.beforeClearStyles([n]).fromTo("transform","translateX(".concat(u,")"),m?"translateX(-100%)":"translateX(100%)");var z=a(v);x.afterAddWrite((function(){"normal"===x.getDirection()&&z.style.setProperty("display","none")}))}else j.fromTo("transform","translateX(".concat(u,")"),"translateX(".concat(p,")")).fromTo(n,1,.8);if(I){var D=l(I).querySelector(".transition-effect");if(D){var M=D.querySelector(".transition-cover"),U=D.querySelector(".transition-shadow"),$=r(),F=r(),G=r();$.addElement(D).beforeStyles({opacity:"1",display:"block"}).afterStyles({opacity:"",display:""}),F.addElement(M).beforeClearStyles([n]).fromTo(n,.1,0),G.addElement(U).beforeClearStyles([n]).fromTo(n,.7,.03),$.addAnimation([F,G]),j.addAnimation([$])}}N.forEach((function(t){var e=r();e.addElement(t);var o=r();o.addElement(t.querySelector("ion-title"));var a=r(),i=t.querySelectorAll("ion-buttons,[menuToggle]"),c=t.closest("ion-header"),s=null==c?void 0:c.classList.contains("header-collapse-condense-inactive"),f=Array.from(i).filter((function(t){var e=t.classList.contains("buttons-collapse");return e&&!s||!e}));a.addElement(f);var y=r(),b=t.querySelectorAll(":scope > *:not(ion-title):not(ion-buttons):not([menuToggle])");b.length>0&&y.addElement(b);var v=r();v.addElement(l(t).querySelector(".toolbar-background"));var S=r(),h=t.querySelector("ion-back-button");if(h&&S.addElement(h),e.addAnimation([o,a,y,S,v]),x.addAnimation(e),S.fromTo(n,.99,0),a.fromTo(n,.99,0),y.fromTo(n,.99,0),g){if(s||o.fromTo("transform","translateX(".concat(u,")"),m?"translateX(-100%)":"translateX(100%)").fromTo(n,.99,0),y.fromTo("transform","translateX(".concat(u,")"),m?"translateX(-100%)":"translateX(100%)"),v.beforeClearStyles([n,"transform"]),(null==c?void 0:c.translucent)?v.fromTo("transform","translateX(0px)",m?"translateX(-100%)":"translateX(100%)"):v.fromTo(n,"var(--opacity)",0),h&&!L){var T=r();T.addElement(l(h).querySelector(".button-text")).fromTo("transform","translateX(".concat(u,")"),"translateX(".concat((m?-124:124)+"px",")")),e.addAnimation(T)}}else s||o.fromTo("transform","translateX(".concat(u,")"),"translateX(".concat(p,")")).fromTo(n,.99,0).afterClearStyles([d,n]),y.fromTo("transform","translateX(".concat(u,")"),"translateX(".concat(p,")")).afterClearStyles([d,n]),S.afterClearStyles([n]),o.afterClearStyles([n]),a.afterClearStyles([n])}))}return x}catch(H){throw H}})),10)}}}))}();
